// src/interfaces/card_interfaces/UnNormalizedCardInterface.ts
export interface UnNormalizedCardInterface {
  _id:string,
  user_id: string;
  zip: number;
  houseNumber: number;
  street: string;
  city: string;
  country: string;
  state: string;
  alt?: string;
  url?: string;
  description: string;
  web: string;
  email: string;
  phone: string;
  subtitle: string;
  title: string;
}